getting things ready for building the queues and workers
