+++ filesystem/file.go
@@ -2,27 +2,17 @@ package filesystem
 
 import "fmt"
 
-// FileSubItem represents a basic file with attributes like name, path, size, and parentID.
-type FileSubItem struct {
+// File represents a basic file with attributes like name, path, size, and parentID.
+type File struct {
 	Name      string
 	Path      string
 	Identifier string
-	ParentID  *string // Using a pointer to represent optional values
-	Size      *int64  // Size is optional, so we use a pointer
+	ParentID  string // Using a pointer to represent optional values
+	Size      int64  // Size is optional, so we use a pointer
+	LastModified string // LastModified is optional, so we use a pointer
 }
 
-// String method to provide a formatted representation of FileSubItem (similar to __repr__)
-func (f FileSubItem) String() string {
-	return fmt.Sprintf("FileSubItem(name=%s, path=%s, size=%v)", f.Name, f.Path, f.Size)
-}
-
-// File represents a file with separate source and destination attributes.
-type File struct {
-	Source      *FileSubItem // Using pointers to allow nil values
-	Destination *FileSubItem
-}
-
-// String method for File struct
+// String method to provide a formatted representation of File (similar to __repr__)
 func (f File) String() string {
-	return fmt.Sprintf("File(Source=%v, Destination=%v)", f.Source, f.Destination)
+	return fmt.Sprintf("File(name=%s, path=%s, size=%v)", f.Name, f.Path, f.Size)
 }
