+++ go.mod
@@ -2,6 +2,6 @@ module github.com/Voltaic314/Data_Migration_Tool
 
 go 1.23.5
 
-require (
-	github.com/mattn/go-sqlite3 v1.14.24
-)
+require github.com/mattn/go-sqlite3 v1.14.24
+
+require golang.org/x/sys v0.30.0
